% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{add_sample_weight}
\alias{add_sample_weight}
\title{Fill in weights using function \code{\link[=fill_in_sample_weight]{fill_in_sample_weight()}} for groups made up of
\code{trip_id} and \code{sample_id}}
\usage{
add_sample_weight(d)
}
\arguments{
\item{d}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}
}
\description{
Fill in weights using function \code{\link[=fill_in_sample_weight]{fill_in_sample_weight()}} for groups made up of
\code{trip_id} and \code{sample_id}
}
