% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{add_weights_by_yr_sex_group}
\alias{add_weights_by_yr_sex_group}
\title{Fill in weights using function \code{\link[=fill_in_weight]{fill_in_weight()}} for groups made up of \code{year} and \code{sex}}
\usage{
add_weights_by_yr_sex_group(d, lw_m, lw_f)
}
\arguments{
\item{d}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{lw_m}{Output from \link{est_lw_params} for Males}

\item{lw_f}{Output from \link{est_lw_params} for Females}
}
\description{
Fill in weights using function \code{\link[=fill_in_weight]{fill_in_weight()}} for groups made up of \code{year} and \code{sex}
}
