% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_plus}
\alias{age_plus}
\title{All ages older than this will be lumped into this age, caled a plus group}
\format{
A character string
}
\usage{
age_plus
}
\description{
All ages older than this will be lumped into this age, caled a plus group
}
\keyword{datasets}
