% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{apply_lw_rel}
\alias{apply_lw_rel}
\title{Apply the LW relationship to the vector of \code{lengths}}
\usage{
apply_lw_rel(lengths, lw_params)
}
\arguments{
\item{lengths}{A vector of lengths}

\item{lw_params}{A vector of two values, alpha and Beta for the LW relationship}
}
\value{
A vector of weights
}
\description{
Apply the LW relationship to the vector of \code{lengths}
}
