% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowtooth.R
\docType{package}
\name{arrowtooth}
\alias{arrowtooth-package}
\alias{arrowtooth}
\title{arrowtooth package}
\description{
See the README on
\href{https://github.com/pbs-assess/arrowtooth#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pbs-assess/arrowtooth}
  \item Report bugs at \url{https://github.com/pbs-assess/arrowtooth/issues}
}

}
\author{
\strong{Maintainer}: Chris Grandin \email{chris.grandin@gmail.com} (\href{https://orcid.org/0000-0002-6746-3486}{ORCID})

}
