% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burnin}
\alias{burnin}
\title{The number of steps at the beginning of the chain to burn-in (discard)}
\format{
A single numeric value
}
\usage{
burnin
}
\description{
The number of steps at the beginning of the chain to burn-in (discard)
}
\keyword{datasets}
