% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-female-props.R
\name{calc_female_props}
\alias{calc_female_props}
\title{Make a tibble in ggplot format with age- or length-based sex proportions
by year}
\usage{
calc_female_props(d, type = c("age", "length"))
}
\arguments{
\item{ac}{Output from \code{\link[gfplot:tidy_comps]{gfplot::tidy_ages_weighted()}}}

\item{lc}{Output from \code{\link[gfplot:tidy_comps]{gfplot::tidy_lengths_weighted()}}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Make a tibble in ggplot format with age- or length-based sex proportions
by year
}
