% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-numbers-at-age.R
\name{calc_iscam_paa}
\alias{calc_iscam_paa}
\title{Fill in a proportions-at-age table with columns required by ISCAM
so that it can be cut/paste directly into the ISCAM data file}
\usage{
calc_iscam_paa(paa, gear_num = 1, digits = 6)
}
\arguments{
\item{paa}{Output from \code{\link[=calc_paa]{calc_paa()}}}

\item{gear_num}{The gear number as required in the ISCAM data file}
}
\value{
a \link{data.frame}
}
\description{
Fill in a proportions-at-age table with columns required by ISCAM
so that it can be cut/paste directly into the ISCAM data file
}
\examples{
\dontrun{
calc_iscam_paa(calc_paa(calc_naa(comm_ft)), gear_num = 1)
calc_iscam_paa(calc_paa(calc_naa(comm_ss)), gear_num = 2)
calc_iscam_paa(calc_paa(calc_naa(survey_samples, "SYN QCS")), gear_num = 3)
calc_iscam_paa(calc_paa(calc_naa(survey_samples, "SYN HS")), gear_num = 5)
calc_iscam_paa(calc_paa(calc_naa(survey_samples, "SYN WCVI")), gear_num = 6)
calc_iscam_paa(calc_paa(calc_naa(survey_samples, "SYN WCHG")), gear_num = 7)
}
}
