% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-maturity.R
\name{calc_maturity}
\alias{calc_maturity}
\title{Calculate the maturity ogive}
\usage{
calc_maturity(d, type = "length")
}
\arguments{
\item{d}{A \link{data.frame} with the structure of that found in the \code{survey_sets}
object of the \link{gfdata} package}

\item{type}{See \code{\link[gfplot:fit_mat_ogive]{gfplot::fit_mat_ogive()}}}
}
\value{
A list of several values including length at 5\%, 50\%, and 95\% maturity
for males and females, and standard error on length at 50\% and 95\%
}
\description{
Calculate the maturity ogive
}
