% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-numbers-at-age.R
\name{calc_naa}
\alias{calc_naa}
\title{Calculate the Numbers-at-age with sample sizes in a table format}
\usage{
calc_naa(d = NULL, survey_abbrev = NULL, start_age = 1, plus_age = NULL)
}
\arguments{
\item{d}{A \link{data.frame}. Either \code{survey_samples} or \code{commercial_sampels} data
from the \link{gfdata} package}

\item{survey_abbrev}{The survey abbreviation. See \link{gfdata} package}

\item{start_age}{The minimum age to include. This is not a minus group, it is an
minimum age to truncate to}

\item{plus_age}{The maximum age to include. This is a plus group and contains
the number of all ages this age and older}
}
\value{
A \link{data.frame} in tabular format with rows being years and columns, ages.
A new column called \code{nsamp} is added. It holds the number of samples all the ages
in each column came from
}
\description{
Calculate the Numbers-at-age with sample sizes in a table format
}
