% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-numbers-at-age.R
\name{calc_paa}
\alias{calc_paa}
\title{Calculate the proportions-at-age for a \link{data.frame} output by
the \code{\link[=calc_naa]{calc_naa()}} function}
\usage{
calc_paa(naa = NULL)
}
\arguments{
\item{naa}{A \link{data.frame} output by the \code{\link[=calc_naa]{calc_naa()}} function}
}
\value{
A \link{data.frame} in the same format as the input \code{naa}, but
proportions of ages in each year. Each row will sum to 1.
}
\description{
Calculate the proportions-at-age for a \link{data.frame} output by
the \code{\link[=calc_naa]{calc_naa()}} function
}
