% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-prop-catch-by-fleet.R
\name{calc_prop_catch_by_fleet}
\alias{calc_prop_catch_by_fleet}
\title{Calculate the proportion of the catch caught in the Freezer trawler sector
for the \code{last_n_years}}
\usage{
calc_prop_catch_by_fleet(
  ft_catch,
  wb_catch,
  end_year = NULL,
  last_n_years = 10
)
}
\arguments{
\item{ft_catch}{Output from \code{\link[=extract_fleet_catch]{extract_fleet_catch()}} for Freezer trawlers}

\item{wb_catch}{Output from \code{\link[=extract_fleet_catch]{extract_fleet_catch()}} for Wet boats}

\item{end_year}{The last year in the mean calculation}

\item{last_n_years}{Number of recent years to use in mean
proportion calculation}
}
\value{
A single value, the proportion men proportion for the
\code{last_n_years} for Freezer trawlers
}
\description{
Calculate the proportion of the catch caught in the Freezer trawler sector
for the \code{last_n_years}
}
\examples{
\dontrun{
ft_catch <- extract_fleet_catch(catch, include = TRUE)
wb_catch <- extract_fleet_catch(catch, include = FALSE)
calc_prop_catch_by_fleet(ft_catch, wb_catch, 10)
}
}
