% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{calc_sample_weights}
\alias{calc_sample_weights}
\title{Calculate the sample weights given the data have many individual weights
If sample weights are not recorded, they will be filled in}
\usage{
calc_sample_weights(d, lw_params)
}
\arguments{
\item{d}{The data frame as output by \code{\link[=convert_length_to_weight_by_sex]{convert_length_to_weight_by_sex()}}}

\item{lw_params}{A vector of two values, alpha and Beta for the LW relationship}
}
\value{
A data frame with the sample weights filled in
}
\description{
Calculate the sample weights given the data have many individual weights
If sample weights are not recorded, they will be filled in
}
