% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sel-for-iscam-from-ss.R
\name{calc_sel_for_iscam}
\alias{calc_sel_for_iscam}
\title{Calculate selectivity parameters from SS values to
iscam values}
\usage{
calc_sel_for_iscam(age_infl = 9, age_95 = 2, min_age = 0, max_age = 25)
}
\arguments{
\item{age_infl}{Parameter 1 from the SS selectivity}

\item{age_95}{Parameter 2 from the SS selectivity}

\item{min_age}{Minimum age to use in calculation}

\item{max_age}{Maximum age to use in calculation}
}
\value{
A vector of two (the age-at=50\% and sd-at-50\% values
for input into an iscam control file)
}
\description{
Calculate selectivity parameters from SS values to
iscam values
}
\details{
The ISCAM and SS platforms have slightly different equations for selectivity
and that means their parameters are slightly different.l This function takes
in parameters from SS (control.ss_new) which are MLE estimates of
selectivity and finds the values of the ISCAM parameters for input into the
ISCAM control file.
}
