% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{calc_sex_props_comm}
\alias{calc_sex_props_comm}
\title{Calculate the proportion of females by year for the commercial fishery}
\usage{
calc_sex_props_comm(dat, lw_params = NA, data_source_name = "default", ...)
}
\arguments{
\item{dat}{A data frame returned by running \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} or
\code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{lw_params}{A two-element list of vectors of alpha, Beta with males first and females second
in the list}

\item{data_source_name}{Name to place in the data frame to describe these data}

\item{...}{Absorb arguments meant for other functions}
}
\value{
A data frame of totals of males and females and proportions of females
}
\description{
Generate weights for all records in \code{dat} that have lengths but no weights associated
Pseudocode:
For each year:
For each trip:
For each sample ID:
Calculate sex-specific mean weights for this sample
Calculate sex-specific catch weight for this sample
EndFor
Calculate sex-specific mean weight for this trip
Calculate sex-specific total catch weight for this trip
EndFor
Calculate the sex-specific mean weights for each quarter (3-months)
Calculate the sex-specific total catch weights for each quarter (3-months) weighted by total catch weight of sampled trips in the quarter
EndFor
}
