% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-data-prep.R
\name{condition_data_prep}
\alias{condition_data_prep}
\title{Prepare data for body condition modelling}
\usage{
condition_data_prep(
  d,
  survey_abbrev = c("SYN QCS", "SYN HS", "SYN WCVI", "SYN WCHG")
)
}
\arguments{
\item{d}{A list, as output from \code{\link[gfdata:get_data]{gfdata::get_data()}}}

\item{survey_abbrev}{A survey abbreviation as found in
\code{d$survey_sets$survey_abbrev}}
}
\value{
A data frame of survey sets
}
\description{
Prepare data for body condition modelling
}
