% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-factor-by-survey.R
\name{condition_factor_by_survey}
\alias{condition_factor_by_survey}
\title{Plot the condition factor fits by survey}
\usage{
condition_factor_by_survey(
  fit_lst,
  grid_fn = "/srv/arrowtooth/arrowtooth-nongit/data/synoptic_grid.rds"
)
}
\arguments{
\item{fit_lst}{A list of a fit from \link{sdmTMB} and a data frame
of the data used in that fit as returned by \code{\link[=condition_fit]{condition_fit()}}}

\item{grid_fn}{The Synoptic grid data frame}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the condition factor fits by survey
}
