% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-fit.R
\name{condition_fit}
\alias{condition_fit}
\title{Run fit to condition model and return the result}
\usage{
condition_fit(d)
}
\arguments{
\item{d}{A data frame, as returned by \code{\link[=condition_data_prep]{condition_data_prep()}}}
}
\value{
A list of two items, \code{fit} which is a fit from \link{sdmTMB} and \code{ds}
which is a data frame of the data used in that fit
}
\description{
Run fit to condition model and return the result
}
