% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-smooth.R
\name{condition_smooth}
\alias{condition_smooth}
\title{Create a plot of the Condition fit incorporating the log depth smoother}
\usage{
condition_smooth(fit_lst)
}
\arguments{
\item{fit_lst}{A list of a fit from \link{sdmTMB} and a data frame
of the data used in that fit as returned by \code{\link[=condition_fit]{condition_fit()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of the Condition fit incorporating the log depth smoother
}
