% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition-tv-depth.R
\name{condition_tv_depth}
\alias{condition_tv_depth}
\title{Create a plot of effect of time-varying third-order polynomial of depth
on condition}
\usage{
condition_tv_depth(
  fit_lst,
  trawl_temp_fn = "/srv/arrowtooth/arrowtooth-nongit/data/trawl_temp.rds"
)
}
\arguments{
\item{fit_lst}{A list of a fit from \link{sdmTMB} and a data frame
of the data used in that fit as returned by \code{\link[=condition_fit]{condition_fit()}}}

\item{trawl_temp_fn}{The trawl temperature data frame RDS filename}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of effect of time-varying third-order polynomial of depth
on condition
}
