% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sel-for-iscam-from-ss.R
\name{conv_sels}
\alias{conv_sels}
\title{Call \code{\link[=calc_sel_for_iscam]{calc_sel_for_iscam()}} for a list of gears}
\usage{
conv_sels()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the gear names and values for the
age-at=50\% and sd-at-50\% parameterrs for each gear
}
\description{
Call \code{\link[=calc_sel_for_iscam]{calc_sel_for_iscam()}} for a list of gears
}
