% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue-extraction.R
\name{cpue_extraction}
\alias{cpue_extraction}
\title{Extract CPUE for Arrowtooth Flounder}
\usage{
cpue_extraction(
  fn_cpue = file.path(drs$nongit_dir, "data", "arrowtooth-cpue-to-2024-10-21.rds"),
  params,
  ret_params = FALSE
)
}
\arguments{
\item{fn_cpue}{The file name of the cpue index data used}

\item{ret_params}{If \code{TRUE}, return \code{params} list instead}
}
\value{
A data frame, or the \code{params} list, if \code{ret_params} is \code{TRUE}
}
\description{
Extract CPUE for Arrowtooth Flounder
}
\details{
Used for the plots in the CPUE appendix in the 2022 assessment
}
