% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crit}
\alias{crit}
\title{Set gradient magnitude convergence criterion to this value}
\format{
A single numeric value
}
\usage{
crit
}
\description{
Set gradient magnitude convergence criterion to this value
}
\keyword{datasets}
