% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{end_catch_yr}
\alias{end_catch_yr}
\title{Catch data ends in this year for this assessment}
\format{
A character string
}
\usage{
end_catch_yr
}
\description{
Catch data ends in this year for this assessment
}
\keyword{datasets}
