% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-numbers-at-age.R
\name{expand_df_by_col}
\alias{expand_df_by_col}
\title{Expand the \code{d} \link{data.frame} to include the values found in \code{vals}}
\usage{
expand_df_by_col(d = NULL, vals = NULL, colname = NULL)
}
\arguments{
\item{d}{A \link{data.frame}}

\item{vals}{A vector of values to expand the \code{d} data frame to.}

\item{colname}{The quoted name of a column in the \link{data.frame} \code{d}}
}
\value{
A \link{data.frame} of the same structure as \code{d} with possibly more rows
}
\description{
Expand the \code{d} \link{data.frame} to include the values found in \code{vals}
}
\details{
If \code{vals} contains any values that are not in the data frame \code{d},
column \code{colname}, one new row will be added for each of them. If all the
values in \code{vals} are already in the data frame, \code{d} will be returned.
NAs will be inserted for non-value columns
}
