% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-age-comps.R
\name{export_age_comps}
\alias{export_age_comps}
\title{Extract the commercial age proportions for pasting into the iSCAM data file}
\usage{
export_age_comps(
  catch_sets,
  samples,
  type = "commercial",
  gear_num = 1,
  surv_series_name = "SYN QCS",
  sex = c("M", "F"),
  write_to_file = TRUE,
  fn_append = "shoreside",
  append = FALSE,
  ...
)
}
\arguments{
\item{catch_sets}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_catch()}}
or \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{samples}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}
or \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{type}{One of 'commercial' or 'survey'. Whichever is set requires
objects \code{catch_sets} and \code{samples} represent that type. See examples.}

\item{gear_num}{Number of gear to be written in output file}

\item{surv_series_name}{Name of a survey to extract.
See the values in \code{survey_abbrev} column of data frame returned by
\code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} for names}

\item{sex}{sex code, "M" = male, "F" = female.
For split sex (full age comps for each sex) use sex = c("M", "F")}

\item{write_to_file}{If \code{TRUE} write the output to the file. If \code{FALSE}
return it}

\item{fn_append}{Text to append to the filename so that different fleets
can be run}

\item{append}{If \code{TRUE}, append the output to the file. If \code{FALSE},
overwrite the file}

\item{...}{Arguments to pass to \code{\link[gfplot:tidy_comps]{gfplot::tidy_ages_weighted()}}}
}
\value{
if \code{write_to_file} is \code{TRUE}, return nothing, else return the data
frame containing the age comps
}
\description{
Extract the commercial age proportions for pasting into the iSCAM data file
}
\examples{
\dontrun{
catch <- gfdata::get_catch("arrowtooth flounder")
commercial_samples <- gfdata::get_commercial_samples("arrowtooth flounder")
catch_ft <- extract_fleet_catch(catch)
catch_nonft <- extract_fleet_catch(catch, include = FALSE)
comm_ft <- extract_fleet_samples(commercial_samples)
comm_nonft <- extract_fleet_samples(commercial_samples, include = FALSE)
extract_age_comps(catch_ft, comm_ft, gear_num = 1, spp_cat_code = 1, month_fishing_starts = 2, day_fishing_starts = 21, append = FALSE)
extract_age_comps(catch_nonft, comm_nonft, gear_num = 2, spp_cat_code = 1, month_fishing_starts = 2, day_fishing_starts = 21, append = TRUE)
extract_age_comps(survey_sets, survey_samples, type = "survey", gear_num = 6, surv_series_name = "SYN WCHG")
}
}
