% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-catch.R
\name{export_catch}
\alias{export_catch}
\title{Create files with Freezer trawler and Shoreside catch in the format required
for direct pasting into the iSCAM data file}
\usage{
export_catch(
  ct,
  areas = c("3[CD]+", "5[ABCDE]+"),
  years = 1996:2021,
  divisor = 1000000,
  digits = 4,
  write_files = TRUE,
  single_fleet = FALSE,
  fns = c(paste0("catch-ft-", Sys.Date(), ".txt"), paste0("catch-ss-", Sys.Date(),
    ".txt")),
  middle_text = c(gear = 1, group = 1, area = 1, sex = 0, type = 1),
  ...
)
}
\arguments{
\item{ct}{Output from \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{areas}{Area list as required by \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}

\item{years}{Years to include in the output}

\item{divisor}{A value to divide all catch values by}

\item{digits}{Number of significant figures to output}

\item{write_files}{Logical. If \code{TRUE}, write the catch tables to files. If
\code{FALSE}, return a list of two data frames}

\item{single_fleet}{Logical. If \code{TRUE}, the catch will be for one fleet,
which contains the sum of all catch for both fleets}

\item{fns}{A vector of two filenames}

\item{middle_text}{A vector of middle column data to insert in the output}

\item{...}{Arguments passed to \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}} and in turn, to
\code{\link[gfplot:plot_catch]{gfplot::set_fishing_year()}}}
}
\value{
Nothing is \code{write_files} is \code{TRUE}. A list of two data frames if
\code{write_files} is \code{FALSE}
}
\description{
Create files with Freezer trawler and Shoreside catch in the format required
for direct pasting into the iSCAM data file
}
