% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-iscam-data.R
\name{export_iscam_data}
\alias{export_iscam_data}
\title{Export all iSCAM data for pasting into the iSCAM data files}
\usage{
export_iscam_data(dat)
}
\arguments{
\item{dat}{Output from \code{\link[gfdata:get_data]{gfdata::cache_pbs_data()}}. Make sure to set the
argument \code{unsorted_only = FALSE} for arrowtooth}
}
\value{
Nothing, creates several files for cutting/pasting into iSCAM
data files
}
\description{
Export all iSCAM data for pasting into the iSCAM data files
}
