% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-mat-lw-age.R
\name{export_mat_lw_age}
\alias{export_mat_lw_age}
\title{Export maturity, length/weight, and age-at-50\% parameters by
creating a file with the output parameters which can be easily
pasted into an iSCAM data file}
\usage{
export_mat_lw_age(
  surv_samples,
  surv_abbrevs = c("SYN QCS", "SYN WCVI", "SYN HS", "SYN WCHG"),
  areas = c("03", "04", "05", "06", "07", "08", "09"),
  maturity_code = 5,
  maturity_convention_code = 4,
  digits = 7,
  write_file = TRUE,
  fn = paste0("vonb-lw-age50-", Sys.Date(), ".txt")
)
}
\arguments{
\item{surv_samples}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{surv_abbrevs}{The survey abbreviations to filter for.
If \code{NULL}, no filtering will occur}

\item{areas}{The areas to filter for. If \code{NULL}, no filtering
will occur}

\item{maturity_code}{The maturity code as found in the GFBio database}

\item{maturity_convention_code}{The maturity convention code as
found in the GFBio database}

\item{digits}{The number of digits each number should have in the
output}

\item{write_file}{Logical. If \code{TRUE}, write the output to the file,
if \code{FALSE}, return the output as an object}

\item{fn}{Tje file name}
}
\value{
If \code{write_file} is \code{FALSE}, a table of the parameter estimates
}
\description{
Export maturity, length/weight, and age-at-50\% parameters by
creating a file with the output parameters which can be easily
pasted into an iSCAM data file
}
