% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-age-comps.R
\name{extract_age_comps}
\alias{extract_age_comps}
\title{Extract the female age proportions for pasting into the iSCAM data file.}
\usage{
extract_age_comps(catch, comm_samples, fn)
}
\arguments{
\item{catch}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{comm_samples}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{fn}{A file name to write the output to. The output is ASCII text written
using \code{\link[utils:write.table]{utils::write.table()}}}
}
\value{
Nothiung. The output is written to the file
}
\description{
Extract the female age proportions for pasting into the iSCAM data file.
}
\examples{
\dontrun
d <- gfdata::get_commercial_samples("arrowtooth flounder")
ct <- gfdata::get_catch("arrowtooth flounder")
extract_age_comps(ct, d) \%>\% print(n = 100)
}
