% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-catch-data.R
\name{extract_catch_data}
\alias{extract_catch_data}
\title{Extract the female catch data for pasting into the iSCAM data file. Be careful, the default filtering values
in \code{\link[=props_comm]{props_comm()}} which are passed through \code{...} have default value which will filter your data
if you do not provide your own values.}
\usage{
extract_catch_data(catch, comm_samples, ...)
}
\arguments{
\item{catch}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{comm_samples}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{...}{Arguments to be passed on to \code{\link[=props_comm]{props_comm()}}}
}
\value{
A data frame you can cut and paste into the iSCAM data file
}
\description{
Extract the female catch data for pasting into the iSCAM data file. Be careful, the default filtering values
in \code{\link[=props_comm]{props_comm()}} which are passed through \code{...} have default value which will filter your data
if you do not provide your own values.
}
\details{
Proportions female are calculated and multiplied by the total catch, which is a sum of all
catch sources (see group_by/summarize clause in ct assignment below).
For years with no samples (missing years), the mean of all other years' proportions are used.
The year range is determined by what is set in \code{\link[=props_comm]{props_comm()}}, which is passed through \code{...}
}
\examples{
\dontrun
d <- gfdata::get_commercial_samples("arrowtooth flounder")
ct <- gfdata::get_catch("arrowtooth flounder")
extract_catch_data(ct, d, species_category = 1, end_year = 2019) \%>\% print(n=100)
}
