% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-fleet-catch.R
\name{extract_fleet_catch}
\alias{extract_fleet_catch}
\title{Extract catch by fleet, where fleet is defined by a vector of vessel names}
\usage{
extract_fleet_catch(
  catch,
  vessel_names = c("VIKING ENTERPRISE", "NORTHERN ALLIANCE", "OSPREY NO 1", "RAW SPIRIT",
    "PACIFIC LEGACY NO. 1", "SUNDEROEY", "VIKING ALLIANCE"),
  include = TRUE
)
}
\arguments{
\item{catch}{Output from \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{vessel_names}{A vector of vessel names to include in the catch if
\code{include} is \code{TRUE},
or exclude if \code{include} is \code{FALSE}}

\item{include}{include or exclude the \code{vessel_names} from the returned
catch data set}
}
\value{
A filtered catch data frame
}
\description{
Extract catch by fleet, where fleet is defined by a vector of vessel names
}
