% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-fleet-samples.R
\name{extract_fleet_samples}
\alias{extract_fleet_samples}
\title{Extract commercial samples by fleet, where fleet is defined by a vector of vessel ids}
\usage{
extract_fleet_samples(
  comm,
  vessel_ids = c(568, 592, 569, 595, 608, 609, 1727),
  include = TRUE
)
}
\arguments{
\item{comm}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{vessel_ids}{A vector of vessel IDs to include in the commercial samples if \code{include} is \code{TRUE},
or exclude if \code{include} is \code{FALSE}}

\item{include}{include or exclude the \code{vessel_ids} from the returned commercial samples data set}
}
\value{
A filtered commercial samples data frame
}
\description{
Extract commercial samples by fleet, where fleet is defined by a vector of vessel ids
}
