% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-geostat-survey.R
\name{extract_geostat_survey}
\alias{extract_geostat_survey}
\title{Extract data from geostat RDS file for input into iSCAM data file}
\usage{
extract_geostat_survey(
  fn,
  decimals = 2,
  gear_num = 1,
  area = 1,
  group = 1,
  sex = 0,
  survey_timing = 0,
  out_fn = file.path(dirname(fn), paste0("geostat-survey-", gear_num, "-",
    as.character(Sys.Date()), ".txt"))
)
}
\arguments{
\item{fn}{The geostat RDS filename}

\item{decimals}{Number of decimal points to round to for index and weight}

\item{gear_num}{The number to put in the gear column of the output data
table}

\item{area}{The area number to put in the area column of the output data
table}

\item{group}{The group number to put in the group column of the output data
table}

\item{sex}{The sex number to put in the sex column of the output data table}

\item{survey_timing}{The survey_timing value to put in the survey_timing
column of the output data table}

\item{out_fn}{The output filename. The file will be created in the same
directory as the input file}
}
\value{
Nothing
}
\description{
Extract data from geostat RDS file for input into iSCAM data file
}
