% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-age-comps.R
\name{extract_survey_age_comps}
\alias{extract_survey_age_comps}
\title{Extract the survey age proportions for pasting into the iSCAM data file.}
\usage{
extract_survey_age_comps(
  surv_series_names = c("SYN QCS", "SYN HS", "SYN WCVI"),
  write_to_file = FALSE,
  ...
)
}
\arguments{
\item{surv_series_names}{See the values in \code{survey_abbrev} column of data frame returned
by \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} for names of surveys to include}

\item{...}{Arguments to pass to \code{\link[=extract_age_comps]{extract_age_comps()}}}
}
\value{

}
\description{
Extract the survey age proportions for pasting into the iSCAM data file.
}
\examples{
\dontrun
survey_sets <- gfdata::get_survey_sets("arrowtooth flounder")
survwy_samples <- gfdata::get_survey_samples("arrowtooth flounder")
extract_survey_age_comps(catch_sets = survey_sets, samples = survey_samples)
}
