% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{fill_in_sample_weight}
\alias{fill_in_sample_weight}
\title{Add \code{sample_weight} column to the data frame by summing specimen weights}
\usage{
fill_in_sample_weight(d)
}
\arguments{
\item{d}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}
}
\description{
Add \code{sample_weight} column to the data frame by summing specimen weights
}
