% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{fill_in_weight}
\alias{fill_in_weight}
\title{Fill in missing weights with calculation based on length and input parameters}
\usage{
fill_in_weight(d, lw_m, lw_f)
}
\arguments{
\item{d}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{lw_m}{Output from \link{est_lw_params} for Males}

\item{lw_f}{Output from \link{est_lw_params} for Females}
}
\description{
Fill in missing weights with calculation based on length and input parameters
}
