% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mohn.R
\name{format_mohn}
\alias{format_mohn}
\title{Create a data.frame from the retrospectives for input into Mohn's
calculations for retrospective patterns}
\usage{
format_mohn(models, quant_name = "sbt_quants")
}
\arguments{
\item{models}{A list of iscam model objects (class \link{mdl_lst_cls})}
}
\value{
A data.frame in the correct format to be input
into \code{\link[icesAdvice:mohn]{icesAdvice::mohn()}}
}
\description{
Create a data.frame from the retrospectives for input into Mohn's
calculations for retrospective patterns
}
