% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-fit-index.R
\name{geostat_fit_index}
\alias{geostat_fit_index}
\title{Fit a Geo statistical spatial model}
\usage{
geostat_fit_index(
  dat,
  species = "arrowtooth flounder",
  grid_fn = "/srv/arrowtooth/arrowtooth-nongit/data/synoptic_grid.rds",
  formula = catch_weight ~ s(log_depth, k = 5),
  family = tweedie(),
  cutoff = 20,
  bias_correct = TRUE,
  anisotropy = FALSE,
  spatiotemporal = "rw",
  silent = FALSE,
  priors = sdmTMBpriors(matern_s = pc_matern(range_gt = 10, sigma_lt = 4), matern_st =
    pc_matern(range_gt = 10, sigma_lt = 3)),
  ...
)
}
\arguments{
\item{dat}{A data frame containing \code{survey_sets} as returned by
\code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{species}{The species common name}

\item{grid_fn}{Filename for the grid data (RDS file)}

\item{formula}{The model formula}

\item{family}{Family of distribution to use for the model fit}

\item{cutoff}{The minimum allowable triangle edge length. See
\code{\link[sdmTMB:make_mesh]{sdmTMB::make_mesh()}}}

\item{anisotropy}{Logical. If \code{TRUE}, allow for spatial correlation that
is directionally dependent. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{spatiotemporal}{Method used to estimate the spatiotemporal
random fields. Can be 'iid', 'ar1', or 'rw'. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{silent}{Silent or include optimization details? Helpful to set
to \code{FALSE} for models that take a while to fit. See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{priors}{Optional penalties/priors via \code{\link[sdmTMB:priors]{sdmTMB::sdmTMBpriors()}},
See \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}

\item{...}{Arguments passed to \code{\link[sdmTMB:sdmTMB]{sdmTMB::sdmTMB()}}}
}
\value{
A list of three elements: sanity, fit and index
}
\description{
Fit a Geo statistical spatial model
}
