% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-fit-indices.R
\name{geostat_fit_indices}
\alias{geostat_fit_indices}
\title{Fit the Geostatistical models for Arrowtooth}
\usage{
geostat_fit_indices(
  dr = "/srv/arrowtooth/arrowtooth-nongit",
  fn = file.path("/srv/arrowtooth/arrowtooth-nongit/data",
    "arrowtooth-flounder-aug11-2022.rds"),
  lst_species = "arrowtooth flounder"
)
}
\arguments{
\item{dr}{Directory name where 'data' directory resides}

\item{fn}{The name of the file containing output from
\code{\link[gfdata:get_data]{gfdata::cache_pbs_data()}}}
}
\value{
A data frame containing index fits
}
\description{
Fit the Geostatistical models for Arrowtooth
}
