% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-plot-biomass-map-bubbles.R
\name{geostat_plot_biomass_map_bubbles}
\alias{geostat_plot_biomass_map_bubbles}
\title{Plot spatial maps showing biomass density, one for each year}
\usage{
geostat_plot_biomass_map_bubbles(
  d,
  deg_offset = c(-0.5, 0.5),
  meters_offset = c(-10, 10),
  utm_zone = 9
)
}
\arguments{
\item{d}{Output from \code{\link[=geostat_prep_data]{geostat_prep_data()}}}

\item{deg_offset}{Degrees offset vector of 2 values for coastline}

\item{meters_offset}{Meters UTM offset for coordinates used in limits of
plot}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot spatial maps showing biomass density, one for each year
}
