% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-plot-individual-vs-design.R
\name{geostat_plot_individual_vs_design}
\alias{geostat_plot_individual_vs_design}
\title{Plot the design-based vs geostatistical survey indices}
\usage{
geostat_plot_individual_vs_design(
  surv_index,
  surv_geostat_dr = "/srv/arrowtooth/arrowtooth-nongit/survey-geostat"
)
}
\arguments{
\item{surv_index}{The data frame returned by \code{\link[gfdata:get_data]{gfdata::get_survey_index()}}}

\item{surv_geostat_dr}{The directory name where the geostatistical model
output RDS files reside e.g. 'i-arrowtooth-flounder-16-no-depth.rds'. The
RDS files themselves must begin with 'i-arrow'}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the design-based vs geostatistical survey indices
}
