% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-plot-map.R
\name{geostat_plot_map}
\alias{geostat_plot_map}
\title{Plot a coastline map for geostat spatial model output}
\usage{
geostat_plot_map(
  d,
  pred,
  column,
  max_color = max(pred[[column]]),
  deg_offset = c(-0.5, 0.5),
  meters_offset = c(-10, 10),
  utm_zone = 9
)
}
\arguments{
\item{d}{A data.frame as output by \code{\link[=geostat_prep_data]{geostat_prep_data()}}}

\item{pred}{A data.frame as output by \code{\link[stats:predict]{stats::predict()}}}

\item{column}{Column name for fill and color}

\item{deg_offset}{Degrees offset vector of 2 values for coastline}

\item{meters_offset}{Meters UTM offset for coordinates used in limits of
plot}

\item{max_colour}{The maximum value for color scale}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a coastline map for geostat spatial model output
}
