% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-plot-stitched-indices.R
\name{geostat_plot_stitched_indices}
\alias{geostat_plot_stitched_indices}
\title{Plot a time series of the various stitched indices}
\usage{
geostat_plot_stitched_indices(
  d,
  fn = "/srv/arrowtooth/arrowtooth-nongit/geostat-figs/geo-delta-gamma-depth.rds"
)
}
\arguments{
\item{d}{Output from \code{\link[=geostat_prep_data]{geostat_prep_data()}}}

\item{fn}{File name for the RDS file containing model outputs}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot a time series of the various stitched indices
}
