% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-prep-data.R
\name{geostat_prep_data}
\alias{geostat_prep_data}
\title{Prepare data for Geostatistical spatial model}
\usage{
geostat_prep_data(
  fn = "/srv/arrowtooth/arrowtooth-nongit/data/arrowtooth-flounder-aug11-2022.rds",
  region = c("QCS", "HS", "WCVI", "WCHG"),
  utm_crs = 32609
)
}
\arguments{
\item{fn}{Filename of the RDS file which includes output list from
\code{\link[gfdata:get_data]{gfdata::cache_pbs_data()}}}

\item{region}{A vector of region abbreviations to include in the data}

\item{utm_crs}{A Universal Transverse Mercator Coordinate Reference System
number}
}
\value{
A data frame containing survey set data with area swept
calculations and UTM coordinates
}
\description{
Prepare data for Geostatistical spatial model
}
