% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-at-fifty-maturity.R
\name{get_age_at_50_mat}
\alias{get_age_at_50_mat}
\title{Title}
\usage{
get_age_at_50_mat(
  surv,
  surv_series = 1:4,
  maturity_code = 5,
  maturity_convention_code = 4,
  start_year = 1996,
  end_year = 2019,
  sex = "both"
)
}
\arguments{
\item{surv}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{surv_series}{A vector of integers representing the survey ids to use
1:4 = c("SYN QCS", "OTHER HS MSA", "SYN HS", "SYN WCVI")}

\item{maturity_code}{cutoff value for maturity code. See table in \code{details} section
This value and above is considered mature and below this value is considered immature}

\item{maturity_convention_code}{Type of maturity scale to use. Flatfish is code 4 (default)}

\item{start_year}{First year to include}

\item{end_year}{Last year to include}

\item{sex}{One of "both", "female" or "male"}
}
\value{
A list of two vectors, one for male and one for female where the vector items are age-at-50\% and sd-at-50\% maturity.
}
\description{
Title
}
\details{
The maturity code table for maturity_convention_code 4 (flatfish):
MATURITY_CODE,SPECIMEN_SEX_CODE,MATURITY_NAME,MATURITY_DESC
1, 1, IMMATURE, "TESTES VERY SMALL, STRING-LIKE AND SOMEWHAT TRANSLUCENT OR PINKISH IN COLOUR"
1, 2, IMMATURE, "OVARIES VERY SMALL, TRANSLUCENT OR PINKISH  AND SOMEWHAT GELATINOUS IN TEXTURE"
2, 1, MATURING, "TESTES ENLARGING, A DISTINCT BULGE EVIDENT BUT STILL TRANSLUCENT OR PINKISH IN COLOUR"
2, 2, MATURING, "OVARIES RELATIVELY SMALL, PINKISH-YELLOW OR CREAM IN COLOUR, GRANULAR IN TEXTURE.  NO DISTINCT EGGS VISIBLE"
3, 1, DEVELOPING, "TESTES ENLARGING, BROWN-WHITE OR WHITE IN COLOUR, FIRM IN TEXTURE"
3, 2, DEVELOPING, "OVARIES LARGE, CREAM OR YELLOW IN COLOUR CONTAINING OPAQUE EGGS THAT CAN BE DISTINGUISED BY DIRECT OBSERVATION.  SEX MAY BE DETERMINED EXTERNALLY"
4, 1, RIPE, "TESTES LARGE, WHITE AND EASILY BROKEN. NO SPERM EVIDENT"
4, 2, GRAVID, "OVARIES CONTAINING PARTLY OR WHOLLY TRANSLUCENT EGGS.  SEX EASILY DETERMINED EXTERNALLY"
5, 1, SPAWNING, "TESTES LARGE, WHITE AND SPERM EVIDENT"
5, 2, RIPE, "OVARIES CONTAINING ENTIRELY TRANSLUCENT, MATURE OVA.  EGGS LOOSE AND WILL RUN FROM OVIDUCTS UNDER SLIGHT PRESSURE"
6, 1, SPENT, "TESTES FLACCID, SHRUNKEN AND YELLOW-BROWN IN COLOUR. SPERM DUCTS ENLARGED AND A SMALL AMOUNT OF SPRM MAY BE PRESENT"
6, 2, SPENT, "OVARIES LARGE, FLACCID AND PURPLE IN COLOUR; A FEW TRANSLUCENT EGGS MAY BE LEFT.  OVARIAN MEMBRANE VERY VASCULAR (BLOODSHOT) AND SAC-LIKE"
7, 1, RESTING, "TESTES FIRM, SMALL AND YELLOW-BROWN IN COLOUR.  SPERM DUCTS SMALL"
7, 2, RESTING, "OVARIES CONTRACTED AND FIRM, PINKISH GREY TO CREAM-YELLOW IN COLOUR AND MAY APPEAR GRANULAR IN TEXTURE BUT NO DISTINCT EGGS ARE VISIBLE"
}
\examples{
\dontrun{
survey_samples <- gfdata::get_survey_samples("arrowtooth flounder")
get_age_at_50_mat(survey_samples)
}
}
