% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bash_path}
\alias{get_bash_path}
\title{Return the location of the bash scripts for this package on your machine}
\usage{
get_bash_path()
}
\value{
The full path where the bash scripts are located
}
\description{
Return the location of the bash scripts for this package on your machine
}
