% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-model.R
\name{get_cvs}
\alias{get_cvs}
\title{Get the lower and upper CV values for a parameter of the length model}
\usage{
get_cvs(vbm, param_name = NULL)
}
\arguments{
\item{vbm}{Output of \code{\link[=calc_vb]{calc_vb()}}}

\item{param_name}{Name of the parameter as found in the \code{param_name} column of \code{vbm}}
}
\value{
A two-element vector of the lower and upper CVs of the parameter
}
\description{
Get the lower and upper CV values for a parameter of the length model
}
