% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsmas}
\alias{hsmas}
\title{Short for Hecate Strait Multispecies Assemblage Survey}
\format{
A character string
}
\usage{
hsmas
}
\description{
Short for Hecate Strait Multispecies Assemblage Survey
}
\keyword{datasets}
