% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_assess_yr}
\alias{last_assess_yr}
\title{The last year the stock was assessed}
\format{
A character string
}
\usage{
last_assess_yr
}
\description{
The last year the stock was assessed
}
\keyword{datasets}
