% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostat-make-grid.R
\name{make_grid}
\alias{make_grid}
\title{Create a grid data frame for the Geo statistical analyses}
\usage{
make_grid(x, years)
}
\arguments{
\item{x}{A data frame with location}

\item{years}{The years to append to replicate rows in the grid data frame}
}
\value{
A data frame
}
\description{
Create a grid data frame for the Geo statistical analyses
}
