% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-re-dat.R
\name{make_re_dat}
\alias{make_re_dat}
\title{Make a data frame from given GLMM model output showing the estimated
parameter values for each parameter}
\usage{
make_re_dat(object)
}
\arguments{
\item{object}{A model as returned by \code{\link[=run_discard_index_model]{run_discard_index_model()}}}
}
\value{
A data frame representing GLMM model parameter value outputs
}
\description{
Make a data frame from given GLMM model output showing the estimated
parameter values for each parameter
}
