% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maxfn}
\alias{maxfn}
\title{The maximum number of function evaluations to allow in the MCMC}
\format{
A single numeric value
}
\usage{
maxfn
}
\description{
The maximum number of function evaluations to allow in the MCMC
}
\keyword{datasets}
