% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_chain_length}
\alias{mcmc_chain_length}
\title{The chain length of the Monte Carlo Markov Chain (MCMC)}
\format{
A single numeric value
}
\usage{
mcmc_chain_length
}
\description{
The chain length of the Monte Carlo Markov Chain (MCMC)
}
\keyword{datasets}
