% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcsave}
\alias{mcsave}
\title{Save the output every nth step of the MCMC chain.
If this is 1, save every step, if 2, every second step, etc.}
\format{
A single numeric value
}
\usage{
mcsave
}
\description{
Save the output every nth step of the MCMC chain.
If this is 1, save every step, if 2, every second step, etc.
}
\keyword{datasets}
