% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcscale}
\alias{mcscale}
\title{Rescale step size for this many initial steps in the MCMC}
\format{
A single numeric value
}
\usage{
mcscale
}
\description{
Rescale step size for this many initial steps in the MCMC
}
\keyword{datasets}
