% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-paths.R
\name{mod_paths}
\alias{mod_paths}
\title{Modify the path element of each of the model RDS files used in the
assessment}
\usage{
mod_paths(drs)
}
\arguments{
\item{drs}{Output from \code{\link[gfiscamutils:set_dirs]{gfiscamutils::set_dirs()}}}
}
\value{
Nothing
}
\description{
Modify the path element of each of the model RDS files used in the
assessment
}
\details{
Used in the case that the models folders are moved from their original
running location. The \code{path} element was set during the building of
the RDS files in that location to their location at that time.
}
