% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-prop-female-iscam.R
\name{modify_prop_female_iscam}
\alias{modify_prop_female_iscam}
\title{Modifies an iSCAM data file to use the proportion female supplied}
\usage{
modify_prop_female_iscam(fn, prop = 0.79)
}
\arguments{
\item{prop}{The proportion (equal or less than 1) of females to insert}

\item{fns}{The file name (*.dat) file to modify}
}
\value{
Nothing
}
\description{
Modifies an iSCAM data file to use the proportion female supplied
}
