% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nox}
\alias{nox}
\title{Don’t show vector and gradient values in function minimizer screen report}
\format{
A logical. TRUE of FALSE
}
\usage{
nox
}
\description{
Don’t show vector and gradient values in function minimizer screen report
}
\keyword{datasets}
