% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{num_posts_used}
\alias{num_posts_used}
\title{The tot_num_posts minus the burnin value. This is the actual number
of chain steps used in the assessment}
\format{
A single numeric value
}
\usage{
num_posts_used
}
\description{
The tot_num_posts minus the burnin value. This is the actual number
of chain steps used in the assessment
}
\keyword{datasets}
