% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_mat_ogives.R
\name{plot_age_mat_ogives}
\alias{plot_age_mat_ogives}
\title{Create a 2-panel plot (one for each sex) of maturity ogives for
several different scenarios}
\usage{
plot_age_mat_ogives(d)
}
\arguments{
\item{d}{A data frame created  in the document code chunk \code{calc-maturities}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a 2-panel plot (one for each sex) of maturity ogives for
several different scenarios
}
