% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fixed-effect-coefs.R
\name{plot_fixed_effect_coefs}
\alias{plot_fixed_effect_coefs}
\title{Plot fixed effect coefficients in the model run by \code{\link[=run_discard_index_model]{run_discard_index_model()}}}
\usage{
plot_fixed_effect_coefs(object)
}
\arguments{
\item{object}{A model as returned by \code{\link[=run_discard_index_model]{run_discard_index_model()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot fixed effect coefficients in the model run by \code{\link[=run_discard_index_model]{run_discard_index_model()}}
}
