% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion-zero-tows.R
\name{plot_proportion_zero_tows}
\alias{plot_proportion_zero_tows}
\title{Plot the proportion of bottom trawls with zero arrowtooth flounder}
\usage{
plot_proportion_zero_tows(
  cpue_fn = file.path("srv", "arrowtooth", "arrowtooth-nongit", "data",
    "cpue-modern.rds"),
  by_area = FALSE
)
}
\arguments{
\item{cpue_fn}{The filename for the CPUE modern file which was extratced
by running the following:
\code{gfdata::get_cpue_index(gear = "bottom trawl", min_cpue_year = 1996)}}

\item{by_area}{Logical. If \code{TRUE}, make the plot grouped by major
statistical area}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot the proportion of bottom trawls with zero arrowtooth flounder
}
