% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-random-intercepts.R
\name{plot_random_intercepts}
\alias{plot_random_intercepts}
\title{Create a plot of random intercepts for the GLMM discard CPUE index model}
\usage{
plot_random_intercepts(object, re_names = c("locality"))
}
\arguments{
\item{object}{A model as returned by \code{\link[=run_discard_index_model]{run_discard_index_model()}}}

\item{re_names}{A vector of names of GLMM model parameters to include
in the plots}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a plot of random intercepts for the GLMM discard CPUE index model
}
