% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-trawl-footprint.R
\name{plot_trawl_footprint}
\alias{plot_trawl_footprint}
\title{Create a map of commercial trawl footprint overlaid by synoptic survey
grids}
\usage{
plot_trawl_footprint(
  tf_rda_fn = system.file("data/trawlfoot.rda", package = "PBSdata"),
  crs_num = 3156,
  utm_zone = 9,
  bath = c(100, 200, 500),
  x_lim = c(122, 920),
  y_lim = c(5343, 6057),
  rotation_angle = 0,
  rotation_center = c(500, 5700),
  show_majorbound = TRUE,
  major_labels = gfplot:::boundary_labels(utm_zone, xmin = x_lim[1]),
  leg_pos = c(0.97, 0.97),
  trawl_foot_fill = "black",
  trawl_foot_alpha = 1,
  survey_alpha = 0.5
)
}
\arguments{
\item{tf_rda_fn}{The name of the trawl footprint RDA file. Get it from here:
https://github.com/pbs-software/pbs-data/tree/master/PBSdata}

\item{crs_num}{The projection to use  See https://epsg.io/}

\item{x_lim}{The limits for the x-axis of the map in the units defined by
\code{crs_num}}

\item{y_lim}{The limits for the y-axis of the map in the units defined by
\code{crs_num}}

\item{leg_pos}{A vector of the x/y position of the legend inside the panel.
Values are between 0 and 1}

\item{trawl_foot_fill}{The color to use to fill the trawl footprint polygons}

\item{trawl_foot_alpha}{The opacity to use for the trawl footprint polygons}

\item{survey_alpha}{The opacity to use for the survey grid lines}

\item{french}{Logical. If \code{TRUE}, use french names and labels in the plot
as necessary}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Create a map of commercial trawl footprint overlaid by synoptic survey
grids
}
