% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-year-locality-interactions.R
\name{plot_year_locality_interactions}
\alias{plot_year_locality_interactions}
\title{Plot year-locality interactions for the GLMM model for Discard CPUE}
\usage{
plot_year_locality_interactions(object)
}
\arguments{
\item{object}{A model as returned by \code{\link[=run_discard_index_model]{run_discard_index_model()}}}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Plot year-locality interactions for the GLMM model for Discard CPUE
}
