% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_all}
\alias{props_all}
\title{Calculate all proportions female (commercial and all surveys)}
\usage{
props_all(comm_samples, ...)
}
\arguments{
\item{comm_samples}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{...}{Arguments to pass to \code{\link[=props_comm]{props_comm()}} and \code{\link[=props_surv]{props_surv()}}}
}
\value{
A data frame containing all female proportions
}
\description{
Calculate all proportions female (commercial and all surveys)
}
