% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_comm}
\alias{props_comm}
\title{Calculate the proportions female for the commercial data}
\usage{
props_comm(
  d,
  areas = c("03", "04", "05", "06", "07", "08", "09"),
  start_year = 1996,
  end_year = 2020,
  species_category = c(0, 1, 2, 3, 4),
  sample_type = c(1, 2, 6, 7, 8),
  gear = c(1, 8),
  data_source_name = "commercial_coastwide",
  ...
)
}
\arguments{
\item{d}{Output from \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{areas}{Vector of area numbers as character strings, eg. "03" is area 3C}

\item{start_year}{Years equal or greater than this and less or equal to \code{end_year} will be returned}

\item{end_year}{Years equal or less than this and greater or equal to \code{start_year} will be returned}

\item{species_category}{Values of \code{species_category_code} to include. Use NA to include all included in data:
SPECIES_CATEGORY_CODE        SPECIES_CATEGORY_DESC
0                                          UNKNOWN
1                                         UNSORTED
2                          SORTED (UNK. CRITERION)
3                                          KEEPERS
4                                        DISCARDED
5                                          REMAINS
6                        LONGLINE - FISH HEAD ONLY
8 LONGLINE/JIG - FISH LOST AT RAIL/LOST AT SURFACE}

\item{sample_type}{Values of \code{sample_type_code} to include. Use NA to include all included in data:
SAMPLE_TYPE_CODE    SAMPLE_TYPE_DESC ROW_VERSION
0                                       UNKNOWN
1                                  TOTAL  CATCH
2                                        RANDOM
4                                      SELECTED
5                                   STRATIFIED
6             RANDOM FROM RANDOMLY ASSIGNED SET
7   RANDOM FROM SET AFTER RANDOMLY ASSIGNED SET
8    RANDOM FROM SET REQUESTED BY VESSEL MASTER
9                            SELECTED JUVENILES
10                               SELECTED ADULTS
11               OTHER - SEE SAMPLE_TYPE_COMMENT
12                    SELECTED FIRST N SPECIMENS
13  SELECTED FIRST N SPECIMENS - WHOLE FISH ONLY
14                       SELECTED TAG RECOVERIES
15                                SELECTED MALES
16                              SELECTED FEMALES
17 UNKNOWN SAMPLE TYPE FOR NMFS TRIENNIAL SURVEY}

\item{gear}{Values of \code{gear_code} to include. Use NA to include all included in data:
GEAR_CODE                                GEAR_DESC
0                                          UNKNOWN
1                                     BOTTOM TRAWL
2                                             TRAP
3                                          GILLNET
4                                         HANDLINE
5                                         LONGLINE
6                                   MIDWATER TRAWL
7                                            TROLL
8                                    UNKNOWN TRAWL
9                                            SEINE
10                                              JIG
11                                     SHRIMP TRAWL
12                          RECREATIONAL ROD & REEL
13                                       PRAWN TRAP
14                                     TUCKER TRAWL
15 MULTI-NET LARGE PLANKTON SAMPLER: HYDROBIOS/MINI
16                                  TWIN BEAM TRAWL
17                                  HAND SEINE 1/2"
18                                PURSE SEINE 3/16"
19                                        BONGO NET
20                                    PROFILING CTD
21                                           CAMERA
22        DROP ACCOUSTIC INFORMATION SYSTEM (DAISY)
23                                  ROSETTE SAMPLER
24                                       ARGO FLOAT
25                                          MOCNESS}

\item{data_source_name}{Name to place in the \code{data_source} column for all records in return data frame}

\item{...}{Arguments not intended for this function}
}
\description{
Calculate the proportions female for the commercial data
}
