% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_data_summary}
\alias{props_data_summary}
\title{Summarize the data used to make the proportions female table}
\usage{
props_data_summary(
  comm_samples,
  surv_samples,
  areas = c("03", "04", "05", "06", "07", "08", "09"),
  start_year = 1996,
  end_year = 2019,
  species_category = c(0, 1, 2, 3, 4),
  sample_type = c(1, 2, 6, 7, 8),
  gear = c(1, 8)
)
}
\arguments{
\item{comm_samples}{Commercial samples as extracted using \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{surv_samples}{Survey samples as extracted using \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}
}
\value{
A data frame summarizing the data
}
\description{
Summarize the data used to make the proportions female table
}
