% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_fishery}
\alias{props_fishery}
\title{Calculate the proportions female for the commercial data}
\usage{
props_fishery()
}
\value{
A data frame containing proportions of females
}
\description{
Calculate the proportions female for the commercial data
}
