% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_surv}
\alias{props_surv}
\title{Calculate the proportions female for the survey data}
\usage{
props_surv(
  surv_series = 1:4,
  surv_series_names = c("qcsss", "hsmas", "hsss", "wcviss"),
  surv_sets = survey_sets,
  surv_samples = survey_samples,
  ...
)
}
\arguments{
\item{surv_series}{A vector of integers representing the survey ids to use}

\item{surv_series_names}{A vector of names of surveys to use for the given \code{surv_series} values}

\item{surv_sets}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{surv_samples}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{...}{Arguments not intended for this function}
}
\value{
A data frame containing proportions of females for the required surveys
}
\description{
Calculate the proportions female for the survey data
}
