% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_survey}
\alias{props_survey}
\title{Calculate the proportions female for the commercial data}
\usage{
props_survey(
  surv_series = 1:4,
  surv_sets = survey_sets,
  surv_samples = survey_samples
)
}
\arguments{
\item{surv_series}{A vector of integers representing the \code{survey_ids} to use}

\item{surv_sets}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{surv_samples}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}
}
\value{
A data frame containing proportions of females
}
\description{
Calculate the proportions female for the commercial data
}
