% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qcsss}
\alias{qcsss}
\title{Short for Queen Charlotte Sound Synoptic Survey}
\format{
A character string
}
\usage{
qcsss
}
\description{
Short for Queen Charlotte Sound Synoptic Survey
}
\keyword{datasets}
