% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-psv-file.R
\name{read_psv_file}
\alias{read_psv_file}
\title{Read in a binary PSV file from an ADMB mcmc output run}
\usage{
read_psv_file(fn, num_records = 1000)
}
\arguments{
\item{fn}{The filename}

\item{num_records}{The number of records to extract from the file.
See \code{\link[=readBin]{readBin()}} for more details}
}
\value{
A matrix with the parameters as coulmns and the records as rows
}
\description{
Read in a binary PSV file from an ADMB mcmc output run
}
