% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_coords}
\alias{rotate_coords}
\title{Rotate vectors of coordinates}
\usage{
rotate_coords(x, y, rotation_angle, rotation_center)
}
\arguments{
\item{x}{The X position vector}

\item{y}{The Y position vector}

\item{rotation_angle}{The angle to rotate the coordinates}

\item{rotation_center}{A vector of 2: X and Y location for the
center to rotate about}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing the rotated coordinates
}
\description{
Rotate vectors of coordinates
}
