% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate_df}
\alias{rotate_df}
\title{Rotate the coordinates in a data frame}
\usage{
rotate_df(df, rotation_angle, rotation_center)
}
\arguments{
\item{df}{An \link{sf} data frame}

\item{rotation_angle}{The angle to rotate the coordinates}

\item{rotation_center}{A vector of 2: X and Y location for the
center to rotate about}
}
\value{
The modified \link{sf} data frame
}
\description{
Rotate the coordinates in a data frame
}
