% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-arrowtooth-sca.R
\name{run_af_sca}
\alias{run_af_sca}
\title{Run the Statistical Catch-at-Age (SCA) model from the \link{MSEtool} package
for Arrowtooth flounder}
\usage{
run_af_sca(main_dirs = set_dirs())
}
\arguments{
\item{main_dirs}{A list of directories used by the project in the format of the
output of the \code{\link[=set_dirs]{set_dirs()}} function}
}
\value{
Open an HTML page in your browser showing a summary of the model
outputs
}
\description{
Run the Statistical Catch-at-Age (SCA) model from the \link{MSEtool} package
for Arrowtooth flounder
}
