% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-discard-index-model.R
\name{run_discard_index_model}
\alias{run_discard_index_model}
\title{Run the discard index GLMM model}
\usage{
run_discard_index_model(
  dfleet,
  params,
  fn_csv = "/srv/arrowtooth/arrowtooth-nongit/data"
)
}
\arguments{
\item{dfleet}{A list, as output from \code{\link[=cpue_extraction]{cpue_extraction()}}}

\item{params}{A parameter list, as returned by \code{\link[=cpue_extraction]{cpue_extraction()}} with its
argument \code{ret_params} set to \code{TRUE}}
}
\value{
The model object, and a CSV is written with the output
}
\description{
Run the discard index GLMM model
}
