% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{start_catch_yr}
\alias{start_catch_yr}
\title{Catch data starts in this year for this assessment}
\format{
A character string
}
\usage{
start_catch_yr
}
\description{
Catch data starts in this year for this assessment
}
\keyword{datasets}
