% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-prop-female.R
\name{table_prop_female}
\alias{table_prop_female}
\title{Create a table of proportions female}
\usage{
table_prop_female(
  prop_lst,
  end_yr = 2019,
  return_df = FALSE,
  format = "latex",
  yrs = NULL,
  ret_means = FALSE,
  bold_headers = TRUE,
  ...
)
}
\arguments{
\item{prop_lst}{A list of two outputs, the first from \code{\link[=props_comm]{props_comm()}}, and
the second from \code{\link[=props_surv]{props_surv()}}}

\item{end_yr}{The last year to include in the table}

\item{return_df}{If \code{TRUE}, return a data .frame instead of a
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{format}{The format of table as in \code{\link[knitr:kable]{knitr::kable()}}}

\item{yrs}{A vector of years to include in the output table. If \code{NULL},
all years will be included}

\item{ret_means}{Logical. If \code{TRUE}, return a list of the gear mean values.
Takes priority over \code{ret_df}}

\item{bold_headers}{If \code{TRUE}, make all column headers bold}

\item{...}{Arguments passed to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of proportions female
}
\details{
If \code{yrs} includes the last year, the mean row will also be included.
\code{yrs} is meant to split up the table for display purposes. The mean
is always calculated from ALL years, even if only some are returned.
}
