% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-prop-female.R
\name{table_prop_female_weights}
\alias{table_prop_female_weights}
\title{Create a table of weights used in the proportion female analysis}
\usage{
table_prop_female_weights(
  samples = NULL,
  type = c("commercial", "survey"),
  end_yr = 2019,
  ord = c("QCS", "HS", "WCVI", "WCHG"),
  return_df = FALSE,
  ...
)
}
\arguments{
\item{samples}{Output from either \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}} or
\code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} depending on the value of \code{type}}

\item{type}{Which type to return, "commercial" or "survey"}

\item{end_yr}{The last year to include in the table
\code{\link[csasdown:csas_table]{csasdown::csas_table()}}}

\item{ord}{A vector of survey names as they appear in the output data frame,
in the order you want them in the output. If \code{NULL}, order will be ignored}

\item{return_df}{If \code{TRUE}, return a data .frame instead of a}

\item{...}{Arguments passed to \code{\link[csasdown:csas_table]{csasdown::csas_table()}}}
}
\value{
A \code{\link[csasdown:csas_table]{csasdown::csas_table()}}
}
\description{
Create a table of weights used in the proportion female analysis
}
