% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tot_num_posts}
\alias{tot_num_posts}
\title{The mcmc_chain_length divded by the mcsave value}
\format{
A single numeric value
}
\usage{
tot_num_posts
}
\description{
The mcmc_chain_length divded by the mcsave value
}
\keyword{datasets}
