% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wcviss}
\alias{wcviss}
\title{Short for West Coast Vancouver Island Synoptic Survey}
\format{
A character string
}
\usage{
wcviss
}
\description{
Short for West Coast Vancouver Island Synoptic Survey
}
\keyword{datasets}
