% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-extra-indices.R
\name{add_extra_indices}
\alias{add_extra_indices}
\title{Add survey indices from files into the survey index object}
\usage{
add_extra_indices(
  surv_index = NULL,
  iphc = NULL,
  discard_cpue = NULL,
  stitched_syn = NULL
)
}
\arguments{
\item{surv_index}{A survey index object as returned by \code{\link[gfdata:get_data]{gfdata::get_survey_index()}}}

\item{iphc}{The IPHC index data}

\item{discard_cpue}{The Discard CPUE index data}

\item{stitched_syn}{The Stitched Synoptics index data}
}
\value{
A survey index object as returned by \code{\link[gfdata:get_data]{gfdata::get_survey_index()}}, with the additional indices added
}
\description{
Add survey indices from files into the survey index object
}
