% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{calc_mean_total_weight_comm}
\alias{calc_mean_total_weight_comm}
\title{Calculate the mean weight and total weight by sex weighted by the sample weight
of each sample in each trip in the commercial fishery}
\usage{
calc_mean_total_weight_comm(d, weight_adj = 1)
}
\arguments{
\item{d}{The data as output from \code{\link[=calc_sample_weights]{calc_sample_weights()}}}

\item{weight_adj}{A value to divide weights by}
}
\value{
A data frame of the values described above
}
\description{
Calculate the mean weight and total weight by sex weighted by the sample weight
of each sample in each trip in the commercial fishery
}
