% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{est_lw_params}
\alias{est_lw_params}
\title{Estimate the length/weight parameters for the dataset \code{dat}}
\usage{
est_lw_params(dat)
}
\arguments{
\item{dat}{A data frame returned by running \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} or
\code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}
}
\value{
The estimated parameters alpha (a) and beta (b) for these data w = a l ^ b or
\code{NA} if there are no data in \code{dat}.
}
\description{
Estimate the length/weight parameters for the dataset \code{dat}
}
