% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-age-comps.R
\name{extract_age_comps}
\alias{extract_age_comps}
\title{Extract the female age proportions for pasting into the iSCAM data file.}
\usage{
extract_age_comps(catch, comm_samples, sex = c("M", "F"))
}
\arguments{
\item{catch}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{comm_samples}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{sex}{sex code, "M" = male, "F" = female.
For split sex (full age comps for each sex) use sex = c("M", "F")}
}
\value{
Nothing. The output is written to the file
}
\description{
Extract the female age proportions for pasting into the iSCAM data file.
}
\examples{
\dontrun
d <- gfdata::get_commercial_samples("arrowtooth flounder")
ct <- gfdata::get_catch("arrowtooth flounder")
extract_age_comps(ct, d) \%>\% print(n = 100)
}
