% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-catch-data.R
\name{extract_catch_data}
\alias{extract_catch_data}
\title{Extract the female catch data for pasting into the iSCAM data file. Be careful, the default filtering values
in \code{\link[=props_comm]{props_comm()}} which are passed through \code{...} have default value which will filter your data
if you do not provide your own values.}
\usage{
extract_catch_data(
  catch,
  comm_samples,
  gear_num = 1,
  append = FALSE,
  female_only = FALSE,
  write_to_file = TRUE,
  extrap_years = NA,
  month_fishing_starts = 1,
  day_fishing_starts = 1,
  ...
)
}
\arguments{
\item{catch}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_catch()}}}

\item{comm_samples}{Data frame as output by \code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{gear_num}{Number of gear to be written in output file}

\item{append}{If \code{TRUE}, append the output to the file. If \code{FALSE}, overwrite the file}

\item{female_only}{If \code{TRUE}, indices and wts will be multiplied by the proportion female}

\item{write_to_file}{If \code{TRUE}, write the output to the file. If \code{FALSE}, return the data frame}

\item{extrap_years}{The number of years to look into the past to extrapolate the last year's catch based
on the last day of catch in the last year. If \code{NA}, do not do an extrapolation of the last year's catch}

\item{...}{Arguments to be passed on to \code{\link[=props_comm]{props_comm()}} and \code{\link[gfplot:plot_catch]{gfplot::tidy_catch()}}}
}
\value{
A data frame you can cut and paste into the iSCAM data file
}
\description{
Extract the female catch data for pasting into the iSCAM data file. Be careful, the default filtering values
in \code{\link[=props_comm]{props_comm()}} which are passed through \code{...} have default value which will filter your data
if you do not provide your own values.
}
\details{
Proportions female are calculated and multiplied by the total catch, which is a sum of all
catch sources (see group_by/summarize clause in ct assignment below).
For years with no samples (missing years), the mean of all other years' proportions are used.
The year range is determined by what is set in \code{\link[=props_comm]{props_comm()}}, which is passed through \code{...}
}
\examples{
\dontrun{
Freezer trawlers:
extract_catch_data(extract_fleet_catch(catch, include = TRUE), commercial_samples, gear_num = 1, append = FALSE,
species_category = 1, month_fishing_starts = 1, day_fishing_starts = 1, female_only = FALSE)
Non-Freezer trawlers:
extract_catch_data(extract_fleet_catch(catch, include = FALSE), commercial_samples, gear_num = 2, append = TRUE,
species_category = 1, month_fishing_starts = 1, day_fishing_starts = 1, female_only = FALSE)
}
}
