% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{make_sex_props}
\alias{make_sex_props}
\title{Create a matrix of proportion female by year from the surveys given in the vector}
\usage{
make_sex_props(dat, areas = 3:9, weight_adj = 1, type = "commercial")
}
\arguments{
\item{dat}{A data frame returned by running \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}} or
\code{\link[gfdata:get_data]{gfdata::get_commercial_samples()}}}

\item{areas}{Management areas to filter data with}

\item{weight_adj}{A value to divide weights by}

\item{type}{One of "commercial" or "survey"}
}
\value{
A data frame of totals of males and females and proportions of females
}
\description{
Create a matrix of proportion female by year from the surveys given in the vector
}
