% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-setup.R
\name{model_setup}
\alias{model_setup}
\title{Load models and set up the directory names}
\usage{
model_setup(main_dirs = NULL, overwrite_rds_files = FALSE)
}
\arguments{
\item{main_dirs}{Output from \code{\link[=set_dirs]{set_dirs()}}}

\item{overwrite_rds_files}{Logical. TRUE to overwrite the model RDS files}
}
\value{
A list of three items, the base_model, the list of bridge models, and
the list of sensitivity models. The two lists are groups of models which are
to be compared with each other in the document. This simplifies plotting and table functions.
}
\description{
Load models and set up the directory names
}
