% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R
\name{props_surv}
\alias{props_surv}
\alias{props_surv_data_summary}
\title{Calculate the proportions female for the survey data}
\usage{
props_surv(
  surv_series = 1:4,
  surv_series_names = c("qcsss", "hsmas", "hsss", "wcviss"),
  surv_sets = survey_sets,
  surv_samples = survey_samples,
  ...
)

props_surv_data_summary(
  surv_samples,
  surv_series = 1:4,
  surv_series_names = c("QCS Synoptic", "HS Multispecies", "HS Synoptic",
    "WCVI Synoptic")
)
}
\arguments{
\item{surv_series}{Values of \code{survey_series_id} which is a column of \code{surv_samples}}

\item{surv_series_names}{Names to be associated to the values of \code{surv_series}}

\item{surv_sets}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{surv_samples}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{...}{Arguments not intended for this function}
}
\value{
A data frame containing proportions of females for the required surveys

A data frame summarizing the data
}
\description{
Calculate the proportions female for the survey data

Summarize the data used to make the proportions female table
}
