% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-sex-props.R, R/extract-survey-indices.R
\name{props_surv}
\alias{props_surv}
\alias{props_surv_data_summary}
\alias{extract_survey_indices}
\title{Calculate the proportions female for the survey data}
\usage{
props_surv(
  surv_series = 1:4,
  surv_series_names = c("qcsss", "hsmas", "hsss", "wcviss"),
  surv_sets = survey_sets,
  surv_samples = survey_samples,
  ...
)

props_surv_data_summary(
  surv_samples,
  surv_series = 1:4,
  surv_series_names = c("QCS Synoptic", "HS Multispecies", "HS Synoptic",
    "WCVI Synoptic")
)

extract_survey_indices(
  survey_index,
  surv_series = c(2, 3, 4, 5, 17),
  surv_series_names = c("SYN QCS", "OTHER HS MSA", "SYN HS", "SYN WCVI", "SYN WCHG"),
  iphc = NULL,
  discard_cpue = NULL,
  stitched_syn = NULL,
  write_to_file = TRUE,
  append = FALSE,
  ...
)
}
\arguments{
\item{surv_series}{Values of \code{survey_series_id} which is a column of \code{surv_samples}}

\item{surv_series_names}{Names to be associated to the values of \code{surv_series}}

\item{surv_sets}{A data frame object returned from \code{\link[gfdata:get_data]{gfdata::get_survey_sets()}}}

\item{surv_samples}{Output from \code{\link[gfdata:get_data]{gfdata::get_survey_samples()}}}

\item{...}{Arguments passed to \code{\link[=props_comm]{props_comm()}}}

\item{survey_index}{Survey index data frame as output by \code{\link[gfdata:get_data]{gfdata::get_survey_index()}}}

\item{iphc}{The IPHC index as read in from iphc-survey-index.rds}

\item{discard_cpue}{The discard CPUE index as read from
cpue-predictions-arrowtooth-flounder-modern-3CD5ABCDE-discard-july-26-feb-fishing-year}

\item{stitched_syn}{The stitched synoptic index as read in from stitched-syn-index.rds}

\item{write_to_file}{If \code{TRUE}, write the output to the file. If \code{FALSE}, return the data frame}

\item{append}{If \code{TRUE}, append the output to the file. If \code{FALSE}, overwrite the file}
}
\value{
A data frame containing proportions of females for the required surveys

A data frame summarizing the data

A list of survey indices for pasting into a iSCAM data file
}
\description{
Calculate the proportions female for the survey data

Summarize the data used to make the proportions female table

Extract the female survey index data for pasting into the iSCAM data file.
The 'Discard CPUE' index must be multiplied by the proportions female from the commerciual fishery
so \code{comm_samples} is required
}
\details{
The CV can be calculated as CV = SD / MEAN or from the Standard error: CV = sqrt(exp(SE^2) - 1)
}
\examples{
\dontrun{
si <- gfdata::get_survey_index("arrowtooth flounder")
extract_survey_indices (survey_index, start_year = 1996, end_year = 2019, species_category = 1)
}
}
