% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-arrowtooth-sca.R
\name{run_af_sca}
\alias{run_af_sca}
\title{Run the Statistical Catch-at-Age (SCA) model from the \link{MSEtool} package
for Arrowtooth flounder}
\usage{
run_af_sca(main_dirs = set_dirs())
}
\value{
Open an HTML page in your browser showing a summary of the model
outputs
}
\description{
Run the Statistical Catch-at-Age (SCA) model from the \link{MSEtool} package
for Arrowtooth flounder
}
